refactor(virtualLive): extract common logic of mc timeline
